# Cache `Hanami::App` middleware stack