@jodosha 2018-02-04
@AlfonsoUceda I tried quickly but didn't found anything that didn't require a lot of testing setup. Skipping for now.